added protection to routes

+++ Project/Endpoints/AdvertisementEndpoints.cs
@@ -1,6 +1,8 @@
 using BlocketChallenge.Models;
 using BlocketChallenge.Services;
-
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
+using BlocketChallenge.Helpers;
 namespace BlocketChallenge.Endpoints;
 
 public static class AdvertisementEndpoints
@@ -24,11 +26,14 @@ public static class AdvertisementEndpoints
             return ad is not null ? Results.Ok(ad) : Results.NotFound();
         });
 
-        //Create Endpoint
-        group.MapPost("/", (Advertisement ad, IAdvertisementService service) =>
+        //Protected: Create Endpoint
+        group.MapPost("/", [Authorize] (Advertisement ad, IAdvertisementService service, ClaimsPrincipal user) =>
         {
             try
             {
+                var userId = user.GetUserID();
+                if (userId is null) return Results.Unauthorized();
+
                 service.Create(ad);
                 return Results.Created($"/api/ads/{ad.Id}", ad);
             }
@@ -38,20 +43,40 @@ public static class AdvertisementEndpoints
             }
         });
 
-        //Update Endpoint
-        group.MapPut("/{id:int}", (int id, Advertisement ad, IAdvertisementService service) =>
+        //Protected: Update Endpoint
+        group.MapPut("/{id:int}", [Authorize] (int id, Advertisement UpdatedAd, IAdvertisementService service, ClaimsPrincipal user) =>
         {
-            if (id != ad.Id)
-                return Results.BadRequest("Id failure.");
+            var userId = user.GetUserID();
+            if (userId is null) return Results.Unauthorized();
 
-            service.Update(ad);
+            if (id != UpdatedAd.Id)
+                return Results.BadRequest("Advertisement Id doesnt match.");
+
+            var existingAd = service.GetAdvertisementById(id);
+            if (existingAd is null) return Results.NotFound();
+
+            if (existingAd.SellerId != userId.Value)
+                return Results.Forbid();
+
+            service.Update(UpdatedAd);
 
             return Results.NoContent();
         });
-        
-        //Delete Endpoint
-        group.MapDelete("/{id:int}", (int id, IAdvertisementService service) =>
+
+        //Protected :Delete Endpoint
+        group.MapDelete("/{id:int}", [Authorize] (int id, IAdvertisementService service, ClaimsPrincipal user) =>
         {
+
+            var userId = user.GetUserID();
+            if (userId is null) return Results.Unauthorized();
+
+            var ad = service.GetAdvertisementById(id);
+            if (ad is null) return Results.NotFound();
+
+            
+            if (ad.SellerId != userId.Value)
+                return Results.Forbid();
+
             service.Delete(id);
             return Results.NoContent();
         });
